поправил ошибки

+++ src/main/java/ru/yandex/practicum/filmorate/Exception/FilmNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.Exception;
 
 public class FilmNotFoundException extends RuntimeException {
-    public FilmNotFoundException(long filmId) {
+    public FilmNotFoundException(Integer filmId) {
         super("Фильм с ID " + filmId + " не найден");
     }
 }
