+++ src/main/java/ru/yandex/practicum/filmorate/Exception/LikeAlreadyExistsException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.Exception;
 
 public class LikeAlreadyExistsException extends RuntimeException {
-    public LikeAlreadyExistsException(Long filmId, Long userId) {
+    public LikeAlreadyExistsException(Integer filmId, Integer userId) {
         super("Пользователь с ID " + userId + " уже поставил лайк фильму с ID " + filmId);
     }
 }
