+++ src/main/java/ru/yandex/practicum/filmorate/Exception/LikeNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.Exception;
 
 public class LikeNotFoundException extends RuntimeException {
-    public LikeNotFoundException(Long filmId, Long userId) {
+    public LikeNotFoundException(Integer filmId, Integer userId) {
         super("Пользователь с ID " + userId + " не ставил лайк фильму с ID " + filmId);
     }
 }
